requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.4
psycopg2==2.7.2
pytz==2017.2
